chore: remove unused images and optimize component imports for landing page
