+++ src/layouts/AuthLayout.tsx
@@ -1,4 +1,5 @@
 import Image from 'next/image';
+import Link from 'next/link';
 import type { ReactNode } from 'react';
 import React from 'react';
 
@@ -34,7 +35,9 @@ const AuthLayout = (props: IAuthLayoutProps) => {
       </div>
       <div className="h-full w-full bg-white p-12 sm:h-fit sm:w-3/4 sm:rounded-[26px] md:w-7/12 xl:h-full xl:p-0">
         <div className="m-auto flex h-full w-full flex-col items-center justify-center gap-8 sm:w-5/6 xl:w-7/12">
-          <Logo />
+          <Link href="/">
+            <Logo />
+          </Link>
           {props.children}
         </div>
       </div>
