+++ src/locales/en.json
@@ -413,7 +413,7 @@
       "title": "Explore",
       "btn1": "See all",
       "btn2": "View more",
-      "btn3": " Explore all Human books",
+      "btn3": " Explore all Hubers",
       "card": {
         "text1": "Topics",
         "text2": "Ratings"
